\name{cwhmath-package}
\alias{cwhmath-package}
\docType{package}
\title{
cwh's mathematical functions
}
\description{
Miscellaneous mathematical functions for general use
}
\details{
\tabular{ll}{
Package: \tab cwhmath\cr
Version: \tab 4.2.0\cr
Date: \tab 2006-01-26\cr
Depends: \tab R (>= 2.0.0)\cr
BundleDescription: \tab Miscellaneous functions for general use\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.wsl.ch/staff/christian.hoffmann/\cr
}

Index:
\preformatted{
FinneyCorr              Finney's correction to log normally distributed
                        data.
adaptlob                Numerically evaluate integral using adaptive
                        rule.
adaptsimstp             Internal mathlib functions
ellipse                 Generate ellipses
is.constant             is.constant
lengths.angle           Lengths of two vectors and angle between them.
my.var                  Correlation, Variance and Covariance (Matrices)
normalize               Normalize vectors
num.ident               Check numerical values for identity
persp2                  Central perspective from 3 to 2 dimensions
                        along specified axis
pointfit                Least squares fit of point clouds, or the
                        Procrustes problem.
rec.prot                Protected division
rotangle                Compute rotation angles from orthogonal matrix.
rotm                    Rotation matrix
rtf                     Rational Transfer Function objects for R
signp                   Sign Function
whole.number            Check an array on whole numbers (x in I).
}
}
\author{
Christian W. Hoffmann <christian.hoffmann@wsl.ch>

Maintainer: Christian W. Hoffmann <christian.hoffmann@wsl.ch>
}
\keyword{ package }
