\name{whole.number}
\alias{whole.number}
\title{Check an array on whole numbers (x in I).}
\description{
  \code{whole.number} checks an array whether it consists of whole numbers only (x in I).
}
\usage{
whole.number(x)
}
\arguments{
  \item{x}{A numerical array.}
}
\value{
	\code{\link{TRUE}}, \code{\link{FALSE}}
}
\author{Bill Venables }
\keyword{data}

